Initally address comments.
