+++ koku/api/settings/tags/mapping/view.py
@@ -15,11 +15,11 @@ from rest_framework.views import APIView
 
 from api.common.pagination import ListPaginator
 from api.common.permissions.settings_access import SettingsAccessPermission
-from api.settings.tags.mapping.query_handler import AddChildQueryHandler
 from api.settings.tags.mapping.query_handler import format_tag_mapping_relationship
 from api.settings.tags.mapping.serializers import AddChildSerializer
 from api.settings.tags.mapping.serializers import EnabledTagKeysSerializer
 from api.settings.tags.mapping.serializers import TagMappingSerializer
+from api.settings.tags.mapping.utils import resummarize_current_month_by_tag_keys
 from api.settings.utils import NonValidatedMultipleChoiceFilter
 from api.settings.utils import SettingsFilter
 from reporting.provider.all.models import EnabledTagKeys
@@ -112,8 +112,11 @@ class SettingsTagMappingChildAddView(APIView):
     def put(self, request):
         serializer = AddChildSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
-        query_handler = AddChildQueryHandler(serializer)
-        query_handler.bulk_create_tag_mappings(request.user.customer.schema_name)
+        parent_row = EnabledTagKeys.objects.get(uuid=serializer.data.get("parent"))
+        children_rows = list(EnabledTagKeys.objects.filter(uuid__in=serializer.data.get("children")))
+        tag_mappings = [TagMapping(parent=parent_row, child=child_row) for child_row in children_rows]
+        TagMapping.objects.bulk_create(tag_mappings)
+        resummarize_current_month_by_tag_keys(children_rows, request.user.customer.schema_name)
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
@@ -122,11 +125,11 @@ class SettingsTagMappingChildRemoveView(APIView):
 
     def put(self, request: Request):
         children_uuids = request.data.get("ids", [])
-        if not EnabledTagKeys.objects.filter(uuid__in=children_uuids).exists():
+        enabled_tags = EnabledTagKeys.objects.filter(uuid__in=children_uuids)
+        if not enabled_tags.exists():
             return Response({"detail": "Invalid children UUIDs."}, status=status.HTTP_400_BAD_REQUEST)
-
         TagMapping.objects.filter(child__in=children_uuids).delete()
-
+        resummarize_current_month_by_tag_keys(list(enabled_tags), request.user.customer.schema_name)
         return Response(status=status.HTTP_204_NO_CONTENT)
 
 
@@ -135,9 +138,9 @@ class SettingsTagMappingParentRemoveView(APIView):
 
     def put(self, request: Request):
         parents_uuid = request.data.get("ids", [])
-        if not EnabledTagKeys.objects.filter(uuid__in=parents_uuid).exists():
+        parent_rows = EnabledTagKeys.objects.filter(uuid__in=parents_uuid)
+        if not parent_rows.exists():
             return Response({"detail": "Invalid parents UUIDs."}, status=status.HTTP_400_BAD_REQUEST)
-
         TagMapping.objects.filter(parent__in=parents_uuid).delete()
-
+        resummarize_current_month_by_tag_keys(list(parent_rows), request.user.customer.schema_name)
         return Response({"detail": "Parents deleted successfully."}, status=status.HTTP_204_NO_CONTENT)
